+++ libs/idp-server-authentication-interactors/src/main/java/org/idp/server/authentication/interactors/email/EmailVerificationChallenge.java
@@ -23,6 +23,7 @@ import java.util.Map;
 import java.util.Objects;
 import org.idp.server.platform.date.SystemDateTime;
 import org.idp.server.platform.json.JsonReadable;
+import org.idp.server.platform.random.OneTimePassword;
 
 public class EmailVerificationChallenge implements Serializable, JsonReadable {
